Separate in-house libraries from rest of code
