+++ mini_fiction/frontend/src/legacy/lib/amajaxify.js
